Unify blueprint URL encoding into a single self-contained helper
compressBlueprint() throws when CompressionStream is unavailable, so
every caller had to duplicate the same plain-base64 fallback: once as
encodeBlueprintFallback() in blueprint-editor-core.js, and again inlined
in main.js's Import handler. Add encodeBlueprintParam() next to
compressBlueprint() to own that fallback in one place, matching the
self-contained encodeBlueprintParam() already used by the sibling
playground repos (omeka-s/nextcloud/facturascripts). Both the Run
button and Import now call it directly; encodeBlueprintFallback is gone.
